refactor: factorize code and include old fields as deprecated
