+++ providers/redis/provider.go
@@ -5,10 +5,8 @@ package redis
 
 import (
 	"context"
-	"time"
 
 	"github.com/redis/go-redis/v9"
-	"github.com/valyala/bytebufferpool"
 )
 
 var all = []byte("*")
@@ -21,6 +19,14 @@ func New(cfg Config) (*Provider, error) {
 
 	if cfg.Logger != nil {
 		redis.SetLogger(cfg.Logger)
+	}
+
+	if cfg.MaxConnAge != 0 {
+		cfg.ConnMaxLifetime = cfg.MaxConnAge
+	}
+
+	if cfg.IdleTimeout != 0 {
+		cfg.ConnMaxIdleTime = cfg.IdleTimeout
 	}
 
 	db := redis.NewClient(&redis.Options{
@@ -67,6 +73,14 @@ func NewFailover(cfg FailoverConfig) (*Provider, error) {
 		redis.SetLogger(cfg.Logger)
 	}
 
+	if cfg.MaxConnAge != 0 {
+		cfg.ConnMaxLifetime = cfg.MaxConnAge
+	}
+
+	if cfg.IdleTimeout != 0 {
+		cfg.ConnMaxIdleTime = cfg.IdleTimeout
+	}
+
 	db := redis.NewFailoverClient(&redis.FailoverOptions{
 		MasterName:       cfg.MasterName,
 		SentinelAddrs:    cfg.SentinelAddrs,
@@ -85,8 +99,8 @@ func NewFailover(cfg FailoverConfig) (*Provider, error) {
 		PoolSize:         cfg.PoolSize,
 		MinIdleConns:     cfg.MinIdleConns,
 		MaxIdleConns:     cfg.MaxIdleConns,
-		ConnMaxIdleTime:  cfg.IdleTimeout,
-		ConnMaxLifetime:  cfg.MaxConnAge,
+		ConnMaxIdleTime:  cfg.ConnMaxIdleTime,
+		ConnMaxLifetime:  cfg.ConnMaxLifetime,
 		PoolTimeout:      cfg.PoolTimeout,
 		TLSConfig:        cfg.TLSConfig,
 	})
@@ -113,6 +127,14 @@ func NewFailoverCluster(cfg FailoverConfig) (*Provider, error) {
 		redis.SetLogger(cfg.Logger)
 	}
 
+	if cfg.MaxConnAge != 0 {
+		cfg.ConnMaxLifetime = cfg.MaxConnAge
+	}
+
+	if cfg.IdleTimeout != 0 {
+		cfg.ConnMaxIdleTime = cfg.IdleTimeout
+	}
+
 	db := redis.NewFailoverClusterClient(&redis.FailoverOptions{
 		MasterName:       cfg.MasterName,
 		SentinelAddrs:    cfg.SentinelAddrs,
@@ -133,8 +155,8 @@ func NewFailoverCluster(cfg FailoverConfig) (*Provider, error) {
 		PoolSize:         cfg.PoolSize,
 		MinIdleConns:     cfg.MinIdleConns,
 		MaxIdleConns:     cfg.MaxIdleConns,
-		ConnMaxIdleTime:  cfg.IdleTimeout,
-		ConnMaxLifetime:  cfg.MaxConnAge,
+		ConnMaxIdleTime:  cfg.ConnMaxIdleTime,
+		ConnMaxLifetime:  cfg.ConnMaxLifetime,
 		PoolTimeout:      cfg.PoolTimeout,
 		TLSConfig:        cfg.TLSConfig,
 	})
@@ -151,19 +173,6 @@ func NewFailoverCluster(cfg FailoverConfig) (*Provider, error) {
 	return p, nil
 }
 
-func (p *Provider) getRedisSessionKey(sessionID []byte) string {
-	key := bytebufferpool.Get()
-	key.SetString(p.keyPrefix)
-	key.WriteString(":")
-	key.Write(sessionID)
-
-	keyStr := key.String()
-
-	bytebufferpool.Put(key)
-
-	return keyStr
-}
-
 // Get returns the data of the given session id
 func (p *Provider) Get(id []byte) ([]byte, error) {
 	key := p.getRedisSessionKey(id)
@@ -175,62 +184,4 @@ func (p *Provider) Get(id []byte) ([]byte, error) {
 
 	return reply, nil
 
-}
-
-// Save saves the session data and expiration from the given session id
-func (p *Provider) Save(id, data []byte, expiration time.Duration) error {
-	key := p.getRedisSessionKey(id)
-
-	return p.db.Set(context.Background(), key, data, expiration).Err()
-}
-
-// Regenerate updates the session id and expiration with the new session id
-// of the the given current session id
-func (p *Provider) Regenerate(id, newID []byte, expiration time.Duration) error {
-	key := p.getRedisSessionKey(id)
-	newKey := p.getRedisSessionKey(newID)
-
-	exists, err := p.db.Exists(context.Background(), key).Result()
-	if err != nil {
-		return err
-	}
-
-	if exists > 0 { // Exist
-		if err = p.db.Rename(context.Background(), key, newKey).Err(); err != nil {
-			return err
-		}
-
-		if err = p.db.Expire(context.Background(), newKey, expiration).Err(); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// Destroy destroys the session from the given id
-func (p *Provider) Destroy(id []byte) error {
-	key := p.getRedisSessionKey(id)
-
-	return p.db.Del(context.Background(), key).Err()
-}
-
-// Count returns the total of stored sessions
-func (p *Provider) Count() int {
-	reply, err := p.db.Keys(context.Background(), p.getRedisSessionKey(all)).Result()
-	if err != nil {
-		return 0
-	}
-
-	return len(reply)
-}
-
-// NeedGC indicates if the GC needs to be run
-func (p *Provider) NeedGC() bool {
-	return false
-}
-
-// GC destroys the expired sessions
-func (p *Provider) GC() error {
-	return nil
 }
